Create XmlBuilder class, split Builder.update_jobs
This is intended to make it easier to generate a list of XmlJob's in different
ways, particularly as part of my experimentation with the use of extensible
Python classes that can be used to construct more interesting CI pipelines than
what is possible using yaml.

Unit tests are currently broken, but I am submitting this now to get early
feedback while working to build out my initial set of job and pipeline
abstractions.

Change-Id: I84f66e7227b12ceb6193d71716a01471ff7f4428