Allow the use of !!null as an `scm` value
This is necessary in the case where global defaults sets a templated `scm` value
used by most jobs but which is undesirable for a minority of jobs and where it
would be inconvenient to copy all the desired settings out of the global
defaults into a custom defaults just for that particular set of jobs. Instead,
this patch makes it possilbe to use the YAML "!!null" value for the scm
attribute.

Change-Id: I201c29215a7090b81e84b9d494bd2cf6ea0370ad